+++ sdk/python/pulumi_pulumiservice/__init__.py
@@ -16,6 +16,7 @@ from .drift_schedule import *
 from .environment import *
 from .environment_rotation_schedule import *
 from .environment_version_tag import *
+from .get_current_user import *
 from .get_insights_account import *
 from .get_insights_accounts import *
 from .get_organization_member import *